Update 03_ads_that_kantar_codes_and_we_dont.R
I am a bit confused about lines 65 and 66.  Wouldn't adding in zeros where there is missing coding mess up the precision and recall.  Also I am not sure of the point of lines 36-43.  I understand that only_K is true when both the cmag and wmp are coded as 1 but then why are these going into folder for ads that kantar codes and we dont?  Should I change the code so that instead of putting in zeros, if there is not a value it should be put in the folder where kantar codes and we dont?